Second commit.
I should probably have committed already a few times! New to this
thing... will get better ;)

I changed a few concepts and added functionality.
Important was the concept of a Timeline-object that carries Keyframes
with points and lines that have minimum "weight". In that way i don't
need to drag around ArrayLists of whole graphs just to get the
coordinates of the stuff to draw on the screen. Also in that
timeline-object the keyframes expand into a smooth animation by adding
interim-frames between the keyframes.

The Importer, Exporter and SimpleGraph classes serve now static methods
- there was no point in them being instantiated for just doing one and
the same thing.
The order of the nodes can now be changed, either in chronological order
or sorting recursively to have the "weight" of the graph on the left,
right, in the middle or away from he middle.
The actual re-ordering itself is actually happening in the Nodes
themselves. Which was a very cool shift of encapsulating and
decentralising (if i may say so?) that i might continue also for the
layouting! The sorting with the weight into and away from the middle
happens with a cool mathy method. I explained it in the comments.

Important was also to reference Nodes directly in Edges and in other
Nodes instead of using a String ID, which then had to be used to run the
findNodeByID method way too many times.
Edges have lost their String-ID because it's not needed, only in the
moment of export one is generated because the graphml-format requires
that. Nodes though still have an ID and a value... maybe i'll streamline
that further.

The supermessy GraphToolkit has now been split up and given its parts
right into the Graph class and the layout-parts into a class for each
layout-type that implement the same interface. Actually the
Layout-classes could just as well be static too - but an interface can't
dictate static methods for some reason. I'll see if an abstract-class
construct makes sense.
Cool is also that you can not only click on nodes to create children on
the very right as in Version 1, but can click on a parent and then drag
the mouse down - where you release will be the place between its
siblings where a child will be added.

In the pipeline:
Radial Layout. Which will most likely result in pushing the
layouting-task into the Nodes directly! As it happened already with the
ordering.
Will see if i understand some other tips and design patterns that i got
told and where that takes me...

I really enjoy this process of keeping to develop towards ever more
elegant solutions that yield object-orientation and encapsulation more
'meaningful'!

Thanks for feedback and tips Markus, Martin, Jonas and Samya!