radial layout + animation speed
A "plain" radial layout is added. Plain as in: not smart yet :) the
nodes don't expand yet into free neighbour-space, they just get fewer
and fewer space the longer their ancestry. Planning to build a
RadialSmartLayout...
Oh and please note the way i am "faking" the levelwise and nodewise
animation at the end of the RadialPlainLayout class :) I couldn't track
when the nodes are doing their placement (because graph knows node but
node does not know graph, which i (currently) think is a good & elegant
architecture) because they do it "decentralized themselves" - so i faked
it afterwards by saving their positions but putting them back to 0,0 and
then position them anew, this time under control and with keyframes
added levelwise and nodewise.

Animation speed (as in thread.sleep(this value)) can now be changed at
any time. If no animation is currently happening the change of that
slider will cause the latest animation to be repeated at the new speed.
If the slider is used during an animation the speed will adapt right
away.