save points as hashmap & AnimationPostproduction
As planned in #8 i got rid of the clunky searchPointByID help-method
through using a HashMap with the node/point IDs as keys.

Giving weight to the conceptual element of caring about the
layouting-animation AFTER the actual layout-computation has been done.
This was present in the radialPlainLayout class before but giving it an
extra-class elevates it's conceptual status and may pave the wave for
adding other "postProduction" layouting like a strict left-to-right one
for instance or a depth-first kind of one that starts from the branch on
the very left and works its way to the right. Or starting in the middle
and working towards both sides simultaneously? Cool things possible i'd
imagine.
Not sure yet if it would be better as a static class or by implementing
an interface like Sanjo did with the layout algorithms...